+ fixed type (template help) + chat: double clicking a nickname while holding shift puts the focus on the input box.
git-svn-id: http://miranda.googlecode.com/svn/trunk@4059 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/tabsrmm/chat/options.c
@@ -551,7 +551,7 @@ BOOL CALLBACK DlgProcOptions1(HWND hwndDlg,UINT uMsg,WPARAM wParam,LPARAM lParam
 		break;
     case WM_COMMAND:
         if (LOWORD(wParam) == IDC_CHAT_ENABLE) {
-            DBWriteContactSettingByte(NULL, "PluginDisable", "chat.dll", IsDlgButtonChecked(hwndDlg, IDC_CHAT_ENABLE) ? 1 : 0);
+            DBWriteContactSettingByte(NULL, "PluginDisable", "chat.dll", (BYTE)(IsDlgButtonChecked(hwndDlg, IDC_CHAT_ENABLE) ? 1 : 0));
             MessageBox(0, TranslateT("You should now immediatly restart Miranda to make this change take effect."), TranslateT("tabSRMM Message"), MB_OK);
         }
 		if (	(LOWORD(wParam) == IDC_NICKROW
